# api.yaml
openapi: 3.0.0
info:
  title: ZenCura Hospital Management System API
  description: API documentation for the ZenCura HMS V2, allowing for patient management, appointment scheduling, and clinical reporting.
  version: 2.0.0
servers:
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/patient/doctors:
    get:
      summary: Search and filter available doctors
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of doctors matching criteria
  /api/patient/appointments:
    post:
      summary: Book a new appointment
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                doctor_id:
                  type: integer
                date:
                  type: string
                  format: date
                time:
                  type: string
      responses:
        '201':
          description: Appointment successfully booked
...
  /api/patient/export:
    post:
      summary: Request asynchronous medical history export
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Export task started (Celery background job)
...